+++ temba/ivr/tests.py
@@ -19,7 +19,8 @@ from temba.contacts.models import Contact
 from temba.flows.models import Flow, FlowRun, ActionLog, FlowStep, FlowRevision
 from temba.msgs.models import Msg, IVR, OUTGOING, PENDING
 from temba.orgs.models import get_current_export_version
-from temba.tests import FlowFileTest, MockTwilioClient, MockRequestValidator, MockResponse
+from temba.tests import FlowFileTest, MockResponse
+from temba.tests.twilio import MockTwilioClient, MockRequestValidator
 from six.moves.urllib.parse import urlparse
 from .clients import IVRException
 from .models import IVRCall
