+++ temba/settings.py.dev
@@ -85,9 +85,3 @@ warnings.filterwarnings('error', r"DateTimeField .* received a naive datetime",
 # Make our sitestatic URL be our static URL on development
 # -----------------------------------------------------------------------------------
 STATIC_URL = '/sitestatic/'
-
-# -----------------------------------------------------------------------------------
-# GoFlow
-# -----------------------------------------------------------------------------------
-FLOW_SERVER_URL = 'http://localhost:8080'
-FLOW_SERVER_DEBUG = False
