+++ temba/tests/__init__.py
@@ -1,3 +1,3 @@
 from __future__ import unicode_literals, absolute_import
 
-from .tests import *  # noqa
+from .base import *  # noqa
